Control.Arrow.Conditional

Documentation

class (Arrow c, Profunctor c) => ArrowCond v c | c -> v where

Arrow based interface to implement conditionals.

Associated Types

type Join (c :: * -> * -> *) x y :: Constraint

Type class constraint used by the abstract instances to join arrow computations.

Methods

if_ :: Join c (x, y) z => c x z -> c y z -> c (v, (x, y)) z

if_ f g -< (v,(x,y)) performs a case distinction on the given value v and executes either (f -< x) or (g -< y). Abstract instances might join the results of f and g.