Control.Arrow.Environment

Documentation

class (Arrow c, Profunctor c) => ArrowEnv var val env c | c -> var, c -> val, c -> env where

Arrow-based interface for interacting with environments.

Associated Types

type Join (c :: * -> * -> *) x y :: Constraint

Type class constraint used by the abstract instances to join arrow computations.

Methods

lookup :: Join c ((val, x), x) y => c (val, x) y -> c x y -> c (var, x) y

Lookup a variable in the current environment. If the environment contains a binding of the variable, the first continuation is called and the second computation otherwise.

getEnv :: c () env

Retrieve the current environment.

extendEnv :: c (var, val, env) env

Extend an environment with a binding.

localEnv :: c x y -> c (env, x) y

Run a computation with a modified environment.

lookup' :: (Join c ((val, var), var) val, Show var, IsString e, ArrowFail e c, ArrowEnv var val env c) => c var val

Simpler version of environment lookup.

lookup'' :: (Join c ((val, var), var) y, Show var, IsString e, ArrowFail e c, ArrowEnv var val env c) => c val y -> c var y

extendEnv' :: ArrowEnv var val env c => c a b -> c (var, val, a) b

Run a computation in an extended environment.

bindings :: (ArrowChoice c, ArrowEnv var val env c) => c ([(var, val)], env) env

Add a list of bindings to the given environment.