Control.Arrow.Except

Documentation

class (Arrow c, Profunctor c) => ArrowExcept e c | c -> e where

Arrow-based interface for exception handling.

Associated Types

type Join (c :: * -> * -> *) x y :: Constraint

Type class constraint used by the abstract instances to join arrow computations.

Methods

throw :: c e a

Opertion that throws an exception that can be handled with catch.

try :: Join c (y, (x, e)) z => c x y -> c y z -> c (x, e) z -> c x z

try f g h executes f, if it succeeds the result is passed to g, if it fails the original input is passed to h.

throw' :: ArrowExcept () c => c a b

Simpler version of throw.

try' :: (Join c (y, (x, e)) z, ArrowExcept e c) => c x y -> c y z -> c x z -> c x z

catch :: (Join c (y, (x, e)) y, ArrowExcept e c) => c x y -> c (x, e) y -> c x y

catch f g handles exceptions thrown in f with g.

catch' :: (Join c (y, (x, e)) y, ArrowExcept e c) => c x y -> c e y -> c x y

Simpler version of catch.

finally :: (Join c ((x, y), (x, e)) y, ArrowExcept e c) => c x y -> c x u -> c x y

finally f g executes g, no matter if f throws an exception.

(<+>) :: (Join c (y, (x, e)) y, ArrowExcept e c) => c x y -> c x y -> c x y

Picks the first computation that does not throw an exception.

tryFirst :: (Join c (y, ((x, [x]), e)) y, ArrowChoice c, ArrowExcept e c) => c x y -> c () y -> c [x] y

tryFirst f g -< l executes f on elements of l until one of them does not throw an exception. In case f throws an exception for all elements of l, g is executed.

success :: ArrowExcept e c => c a a

A computation that always succeeds