Control.Arrow.Store

Documentation

class (Arrow c, Profunctor c) => ArrowStore var val c | c -> var, c -> val where

Arrow-based interface to describe computations that read from a store. The parameter y needs to be exposed, because abstract instances may need to join on y.

Associated Types

type Join (c :: * -> * -> *) x y :: Constraint

Methods

read :: Join c ((val, x), x) y => c (val, x) y -> c x y -> c (var, x) y

Reads a value from the store. Fails if the binding is not in the current store.

write :: c (var, val) ()

Writes a value to the store.

read' :: (Show var, Join c ((val, var), var) val, IsString e, ArrowFail e c, ArrowStore var val c) => c var val

Simpler version of read