Control.Arrow.Transformer.Abstract.BoundedEnvironment

Documentation

data EnvT var addr val c x y

Abstract domain for environments in which concrete environments are approximated by a mapping from variables to addresses and a mapping from addresses to values. The number of allocated addresses allows to tune the precision and performance of the analysis.

Abstract domain for environments in which concrete environments are approximated by a mapping from variables to addresses and a mapping from addresses to values. The number of allocated addresses allows to tune the precision and performance of the analysis.

Furthermore, closures and environments are defined mutually recursively. By only allowing a finite number of addresses, the abstract domain of closures and environments becomes finite.

runEnvT :: (Show var, Identifiable var, Identifiable addr, Complete val, ArrowJoin c, ArrowChoice c, Profunctor c) => c (var, val, Map var addr val) addr -> EnvT var addr val c x y -> c ([(var, val)], x) y