Control.Arrow.Transformer.Abstract.Contour

Documentation

data CallString lab

data ContourT lab c a b

Records the k-bounded call string. Meant to be used in conjunction with BoundedEnvironment.

runContourT :: (Arrow c, Profunctor c) => Int -> ContourT lab c a b -> c a b

Runs a computation that records a call string. The argument k specifies the maximum length of a call string. All larger call strings are truncated to at most k elements.