Control.Arrow.Transformer.Concrete.Store

Documentation

newtype StoreT var val c x y

Arrow transformer that adds a store to a computation.

Constructors

StoreT (StateT (HashMap var val) c x y)

runStoreT :: StoreT var val c x y -> c (HashMap var val, x) (HashMap var val, y)

Execute a computation and only return the result value and store.

evalStoreT :: Arrow c => StoreT var val c x y -> c (HashMap var val, x) y

Execute a computation and only return the result value.

execStoreT :: Arrow c => StoreT var val c x y -> c (HashMap var val, x) (HashMap var val)

Execute a computation and only return the result store.