Control.Arrow.Utils

Documentation

map :: ArrowChoice c => c x y -> c [x] [y]

Applies a computation to all elements of the input list and collects the results in an list.

void :: Profunctor c => c x y -> c x ()

Throws away the result of a computation.

(&&>) :: (Arrow c, Profunctor c) => c a () -> c a b -> c a b infixr 1

pi1 :: Arrow c => c (x, y) x

Projects the first component of a product.

pi2 :: Arrow c => c (x, y) y

Projects the second component of a product.

zipWith :: (ArrowChoice c, Profunctor c) => c (x, y) z -> c ([x], [y]) [z]

Zips two lists together.

fold :: ArrowChoice c => c (a, x) a -> c ([x], a) a

Folds a computation over a list from left to right.

duplicate :: Arrow c => c x (x, x)

Duplicates the current value.

const :: Profunctor c => c () x -> c y x

creates a computation that always returns the same value.