Data.Abstract.Except

Documentation

data Except e x

Abstrat domain for exceptions. This abstract domain approximates error more precisely because 'Success ⋢ Fail'. Use this type for analysis in languages that can handle exceptions.

Constructors

Success x

Fail e

SuccessOrFail e x

widening :: Widening e -> Widening a -> Widening (Except e a)

toEither :: Except e a -> Either e (Either a (e, a))

fromMaybe :: Maybe a -> Except () a

toMaybe :: Except e a -> Maybe a