Data.Abstract.IterationStrategy

Documentation

type IterationStrategy c a b = c a b -> c a b

The iteration strategy ▽ ensures that a recursive computation terminates, soundly approximates the fixed point and avoids redundant computation.

It ensures termination by transforming an infinite series x1, x2, x3 ..., to a finite series: x1 ▽ s1 = (Compute x1',s2) & x1 ⊑ x1' x2 ▽ s2 = (Compute x2',s3) & x2 ⊑ x2' x3 ▽ s3 = (Compute x3',s4) & x3 ⊑ x3' ... xn ▽ sn = (Cached y,sn+1) Of course this is only sound if we iterate on inputs that recursively depend on themselves. For example, x1 ▽ s1 = (ComputeAndIterate x1',s2) x2 ▽ s2 = (Compute x2',s3) x3 ▽ s3 = (Cached (lookupDefault x1' bottom s3),s4) Because x3 ⊑ x1', we can reuse the result of x1' at the loss of precision. However, since the result of x1' has not been computed yet, we initialize the result with bottom and iterate on it until it stabilized.

The iteration strategy can also avoid redundant computation by returning cached results instead of recomputing them.

class ArrowRun t where

Methods

run :: Profunctor c => t c a b -> c a b

filter :: (Profunctor c, ArrowChoice c, ArrowApply c) => Prism' a a' -> IterationStrategy c a' b -> IterationStrategy c a b

data Cache a b

Constructors

Cache

Fields

old :: HashMap a b

new :: HashMap a b

newtype ParallelT s a b c x y

Constructors

ParallelT (StateT (Cache a b) (ReaderT (Stack a, s a) c) x y)

parallel :: (Identifiable a, LowerBounded b, Profunctor c, ArrowChoice c) => StackWidening s a -> Widening b -> IterationStrategy (ParallelT s a b c) a b

data Component a

Constructors

Component

Fields

head :: HashSet a

body :: HashSet a

newtype ChaoticT s a b c x y

Constructors

ChaoticT (WriterT (Component a) (StateT (HashMap a (b, Stable)) (ReaderT (Stack a, s a) c)) x y)

runChaoticT :: (IsEmpty (s a), Profunctor c) => ChaoticT s a b c x y -> c x (HashMap a (b, Stable), y)

chaotic :: (Identifiable a, LowerBounded b, Profunctor c, ArrowChoice c, ArrowApply c) => StackWidening s a -> Widening b -> IterationStrategy (ChaoticT s a b c) a b