Data.Abstract.Map

Documentation

data Map a b

Abstract hashmap

singleton :: Identifiable a => a -> b -> Map a b

empty :: Map a b

lookup :: Identifiable a => a -> Map a b -> Maybe b

unsafeLookup :: Identifiable a => a -> Map a b -> Maybe b

insert :: Identifiable a => a -> b -> Map a b -> Map a b

insertWith :: (Identifiable a, Complete b) => (b -> b -> b) -> a -> b -> Map a b -> Map a b

unsafeInsertWith :: Identifiable a => (b -> b -> b) -> a -> b -> Map a b -> Map a b

unsafeInsertWithLookup :: Identifiable a => (b -> b -> b) -> a -> b -> Map a b -> (b, Map a b)

delete :: Identifiable a => a -> Map a b -> Map a b

delete' :: (Foldable f, Identifiable a) => f a -> Map a b -> Map a b

union :: (Identifiable a, Complete b) => Map a b -> Map a b -> Map a b

adjust :: Identifiable a => (b -> b) -> a -> Map a b -> Map a b

toList :: IsList l => l -> [Item l]

The toList function extracts a list of Item l from the structure l. It should satisfy fromList . toList = id.

fromList :: IsList l => [Item l] -> l

The fromList function constructs the structure l from the given list of Item l

mapMaybe :: (Identifiable a', Complete b') => ((a, b) -> Maybe (a', b')) -> Map a b -> Map a' b'

map :: (b -> b') -> Map a b -> Map a b'

compose :: (Identifiable a, Identifiable b, Complete c) => [(a, b)] -> Map b c -> Map a c

widening :: Identifiable a => Widening b -> Widening (Map a b)

fromThereList :: Identifiable a => [(a, (There, b))] -> Map a b