Data.Abstract.StackWidening

Documentation

type StackWidening stack a = a -> stack a -> (a, stack a)

A stack widening operator (▽ :: s -> a -> (s,a)) follows the same idea as a regular widening operator, but does not have the restriction that the input chain has to be ascending: Given an infinite series x1, x2, x3, x4 ..., the stack widening operator produces a series x1 ▽ s1 = (x1',s2) & x1 ⊑ x1' x2 ▽ s2 = (x2',s3) & x2 ⊑ x2' x3 ▽ s3 = (x3',s4) & x3 ⊑ x3' x4 ▽ s4 = (x4',s5) & x4 ⊑ x4' ... such that x1', x2', x3' ... repeats itself, i.e. there exists n,m with n /= m and xn' = xm'.

finite :: StackWidening Unit a

groupBy :: (IsEmpty (stack a'), Identifiable k) => Iso' a (k, a') -> StackWidening stack a' -> StackWidening (Groups stack k a') a

maxSize :: Int -> StackWidening s a -> StackWidening (Size ** s) a

fromWidening :: Complete a => Widening a -> StackWidening s a -> StackWidening (Last ** s) a

reuse :: Identifiable a => (a -> HashSet a -> Maybe a) -> StackWidening s a -> StackWidening (Stack ** s) a

reuseFirst :: (Identifiable a, PreOrd a) => StackWidening s a -> StackWidening (Stack ** s) a

data Groups s k a' a

Constructors

Groups (HashMap k (s a'))

data Size a

Constructors

Size Int

data Stack a

Constructors

Stack (HashSet a)

newtype Last a

Constructors

Last (Maybe a)

data (s1 ** s2) a

Constructors

Product (s1 a) (s2 a)

newtype Const s a' a

Constructors

Const

Fields

getConst :: s a'

data Unit a

Constructors

Unit