Data.Abstract.There

Documentation

data There

Datatype that indicates if a value in the map must be there or may not be there.

Constructors

Must

May

widening :: Widening There

may :: a -> (There, a)

must :: a -> (There, a)