Data.Abstract.Widening

Documentation

type Widening a = a -> a -> (Stable, a)

▽ has to be an upper bound operator, i.e. x ⊑ x ▽ y and y ⊑ x ▽ y. Furthermore, iterating ▽ on an ascending chain has to stabilize: Let x1, x2, ... xn be an infinite ascending chain, then x1, x1 ▽ x2, (x1 ▽ x2) ▽ x3, ... (similar to left fold) has a limit:

... finite ▽ / ▽ / ▽ / x1 ⊑ x2 ⊑ x3 ⊑ x4 ⊑ ... infinite

Furthermore, we allow our widening operators to maintain some state between each iteration.

data Stable

Datatype that signals if the ascending chain stabilized.

Constructors

Stable

Instable

finite :: Complete a => Widening a

For a preorder with no infinite ascending chains, (⊔) is a trivial widening operator.

toJoin :: (Widening a -> Widening b) -> (a -> a -> a) -> b -> b -> b

toJoin2 :: (Widening a -> Widening b -> Widening c) -> (a -> a -> a) -> (b -> b -> b) -> c -> c -> c

bounded :: Complete a => a -> Widening a -> Widening a

Widening operator that joins until the given limit is reached. Then it calls the fallback widening operator.

(**) :: Widening a -> Widening b -> Widening (a, b)