Data.GaloisConnection

Documentation

class (PreOrd x, PreOrd y) => Galois x y where

A galois connection consisting of an abstraction function alpha and a concretization function gamma between two pre-ordered sets has to satisfy forall x,y. alpha x ⊑ y iff x ⊑ gamma y

Methods

alpha :: x -> y

gamma :: y -> x

alphaSing :: Galois (Pow x) x' => x -> x'

lifted :: Complete y => (x -> y) -> Pow x -> y