Data.OrdMap

Documentation

data OrdMap n1 n2

Datatype that describes a subset of a heterogenous order with fast access to the lower and upper sets of elements.

empty :: OrdMap n1 n2

insert :: (Identifiable n1, Identifiable n2) => n1 -> n2 -> Ordering -> OrdMap n1 n2 -> OrdMap n1 n2

compare :: (Identifiable n1, Identifiable n2) => n1 -> n2 -> Ordering -> OrdMap n1 n2 -> Bool

data Ordering

Constructors

LessThanEquals

Equivalent

GreaterThanEquals

NotLessThanEquals

Incomparable

NotGreaterThanEquals

StrictLessThan

StrictGreaterThan